Добавляет задание 5
